requirements.txt
numpy
scipy
paho-mqtt